chore: position of filter is updated
