+++ frontend/src/pages/Logs/styles.ts
